feat: Validate explain format
